uncasted controlnet to float16

+++ src/train_controlnet_severity_ddp.py
@@ -48,7 +48,7 @@ def main(args):
     tokenizer = CLIPTokenizer.from_pretrained("openai/clip-vit-large-patch14")
     text_encoder = CLIPTextModel.from_pretrained("openai/clip-vit-large-patch14").to(device)
     unet    = UNet2DConditionModel.from_pretrained("runwayml/stable-diffusion-v1-5", subfolder="unet").to(device)
-    controlnet = ControlNetModel.from_pretrained("lllyasviel/sd-controlnet-depth", torch_dtype=torch.float16).to(device)
+    controlnet = ControlNetModel.from_pretrained("lllyasviel/sd-controlnet-depth").to(device)
     scheduler = DDPMScheduler.from_pretrained("runwayml/stable-diffusion-v1-5", subfolder="scheduler")
 
     for p in vae.parameters():         p.requires_grad = False
